+++ Netable/Netable/Netable.swift
@@ -113,7 +113,6 @@ open class Netable {
                 }
 
                 guard let response = response as? HTTPURLResponse else { fatalError("Casting response to HTTPURLResponse failed") }
-
                 guard 200...299 ~= response.statusCode else {
                     self.logDestination.log(event: .requestCompleted(statusCode: response.statusCode, responseData: data, finalizedResult: nil))
                     throw NetableError.httpError(response.statusCode, data)